requirements.txt
Flask
python-dotenv
pyodbc
mysql-connector-python
openai
qdrant-client